Melhora no envio de dados e caminho de itens para o front end
