+++ src/main/java/telas/persons/PersonsData.java
@@ -1,0 +1,122 @@
+package src.main.java.telas.persons;
+
+public class PersonsData {
+
+    public PersonsData() {}
+
+    String Curupira = "src/resources/Imagens/Personagens/Curupira.png";
+    public String getCurupira() {
+        return Curupira;
+    }
+
+    public void setCurupira(String curupira) {
+        Curupira = curupira;
+    }
+
+    //////////////////////////////////////
+
+    String CurupiraSobre = "O Curupira é uma figura icônica do folclore brasileiro. Ele é o protetor das florestas, descrito como um ser de\n" +
+            "estatura baixa. Possui cabelos vermelhos e pés virados para trás, o que lhe confere um aspecto singular. Suas\n" +
+            "travessuras são conhecidas por confundir caçadores e desorientar aqueles que tentam invadir o seu território.\n" +
+            "O Curupira é um guardião da natureza, usando truques e armadilhas para proteger os animais e as plantas da\n" +
+            "floresta. Ele se comunica com os seres da mata, e seu assovio ecoa, reverberando entre as árvores, como um\n" +
+            "aviso para os invasores. Apesar de sua aparência assustadora, o Curupira é justo e ajuda aqueles que respeitam\n" +
+            "a floresta e suas criaturas. Ele é uma representação da sabedoria e da resistência da natureza, sempre pronto\n" +
+            "para defender seu lar e garantir que os ecossistemas continuem a florir. O Curupira é um símbolo da luta\n" +
+            "contra a destruição ambiental e a exploração desenfreada, e seu legado continua a inspirar as novas gerações\n" +
+            "a valorizar e proteger a biodiversidade do Brasil.";
+
+    public String getCurupiraSobre() {
+        return CurupiraSobre;
+    }
+
+    public void setCurupiraSobre(String curupiraSobre) {
+        CurupiraSobre = curupiraSobre;
+    }
+
+    /////////////////////////////////////////
+
+    String Saci = "src/resources/Imagens/Personagens/Saci.png";
+    public String getSaci() {
+        return Saci;
+    }
+
+    public void setSaci(String saci) {
+        Saci = saci;
+    }
+
+    /////////////////////////////////////////////////////
+
+    String SaciSobre = "O Saci-Pererê é uma das figuras mais travessas e encantadoras do folclore brasileiro. Ele é representado como um\n" +
+            "menino negro, de uma perna só, que usa um gorro vermelho mágico. Famoso por suas brincadeiras e travessuras, o Saci\n" +
+            "aparece em redemoinhos de vento, pregando peças nos desavisados. Ele esconde objetos, atrapalha tarefas domésticas\n" +
+            "e prega peças em viajantes. Apesar de suas travessuras, o Saci não é um ser mal-intencionado. Ele gosta de\n" +
+            "brincadeiras, mas também ajuda aqueles que o tratam com respeito. A lenda do Saci ensina sobre o valor da\n" +
+            "esperteza e da malícia, além de destacar o respeito pelas forças da natureza. Com seu gorro vermelho, o Saci\n" +
+            "mantém o controle de suas travessuras e guarda segredos da mata, sendo um personagem amado e temido ao mesmo tempo.";
+    public String getSaciSobre() {
+        return SaciSobre;
+    }
+
+    public void setSaciSobre(String saciSobre) {
+        SaciSobre = saciSobre;
+    }
+
+    //////////////////////////////////////////
+
+    String Iara = "src/resources/Imagens/Personagens/Iara.png";
+    public String getIara() {
+        return Iara;
+    }
+
+    public void setIara(String iara) {
+        Iara = iara;
+    }
+
+    ///////////////////////////////////////////////////
+
+    String IaraSobre = "A Iara, também conhecida como Mãe d'Água, é uma das mais encantadoras figuras do folclore brasileiro. Ela é uma\n" +
+            "serena e bela sereia que vive nas águas doces dos rios e lagos da Amazônia. Sua aparência é descrita como a de uma\n" +
+            "mulher de longos cabelos negros e olhos penetrantes, com metade do corpo humano e a outra metade de peixe. Ela é\n" +
+            "famosa por seu canto hipnotizante, que atrai os navegantes para as profundezas das águas. Aqueles que ouvem sua voz\n" +
+            "melodiosa não conseguem resistir ao seu encanto e acabam sendo levados para o fundo dos rios. Apesar de sua beleza\n" +
+            "e fascinação, a Iara pode ser perigosa para aqueles que se aproximam de seu território sem respeito. No entanto,\n" +
+            "algumas histórias contam que a Iara é uma protetora das águas e das criaturas que nela habitam, punindo aqueles\n" +
+            "que poluem ou ameaçam seu ambiente. Assim, ela é uma representação da força misteriosa da natureza e da importância\n" +
+            "de preservar os rios e as florestas do Brasil.";
+    public String getIaraSobre() {
+        return IaraSobre;
+    }
+
+    public void setIaraSobre(String iaraSobre) {
+        IaraSobre = iaraSobre;
+    }
+    ////////////////////////////////////////////////////////////
+
+    String Boitata = "src/resources/Imagens/Personagens/Boitata.gif";
+    public String getBoitata() {
+        return Boitata;
+    }
+
+    public void setBoitata(String boitata) {
+        Boitata = boitata;
+    }
+
+    //////////////////////////////////////////////////////////
+
+    String BoitataSobre = "O Boitatá é uma lenda do folclore brasileiro, conhecido como uma serpente de fogo que protege as florestas e campos\n" +
+            "dos invasores. Ele é descrito como um espírito brilhante que percorre os campos à noite, assustando aqueles que\n" +
+            "tentam destruir a natureza com incêndios. Seu corpo luminoso serpenteia pelo ar, criando um espetáculo assustador\n" +
+            "e fascinante. O Boitatá não é uma criatura maligna, mas sim um defensor feroz do meio ambiente. Ele persegue\n" +
+            "aqueles que ateiam fogo à mata, devolvendo as chamas com grande intensidade. Sua luz ofuscante é tanto uma\n" +
+            "advertência quanto uma punição para os que desrespeitam a natureza. Assim como o Curupira, o Boitatá é um símbolo\n" +
+            "de resistência, uma manifestação da fúria da terra contra a destruição e a falta de respeito pelo equilíbrio\n" +
+            "ambiental.";
+    public String getBoitataSobre() {
+        return BoitataSobre;
+    }
+
+    public void setBoitataSobre(String boitataSobre) {
+        BoitataSobre = boitataSobre;
+    }
+}
