--- src/main/java/telas/persons/PersonsImages.java
@@ -1,46 +1,0 @@
-package src.main.java.telas.persons;
-
-public class PersonsImages {
-
-    public PersonsImages() {}
-
-    String Curupira = "src/resources/Imagens/Personagens/Curupira.png";
-
-    String Saci = "src/resources/Imagens/Personagens/Saci.png";
-
-    String Iara = "src/resources/Imagens/Personagens/Iara.png";
-
-    String Boitata = "src/resources/Imagens/Personagens/Boitata.gif";
-
-    public String getCurupira() {
-        return Curupira;
-    }
-
-    public void setCurupira(String curupira) {
-        Curupira = curupira;
-    }
-
-    public String getSaci() {
-        return Saci;
-    }
-
-    public void setSaci(String saci) {
-        Saci = saci;
-    }
-
-    public String getIara() {
-        return Iara;
-    }
-
-    public void setIara(String iara) {
-        Iara = iara;
-    }
-
-    public String getBoitata() {
-        return Boitata;
-    }
-
-    public void setBoitata(String boitata) {
-        Boitata = boitata;
-    }
-}
