+++ src/resources/Styles/TelaInicial.css
@@ -33,11 +33,11 @@
         -fx-effect: dropshadow(gaussian, black, 5, 0, 0, 2);
 }
 
-#iniciar-button: hover{
+#iniciar-button:hover{
     -fx-background-color: lightblue;
 }
 
-#sobre-button: hover{
+#sobre-button:hover{
     -fx-background-color: lightblue;
 }
